// src/types/updateEmail.ts
export interface updateEmail {
  full_name: string;
  address_line_1: string;
  address_line_2: string;
  postal: string;
  city: string;
  province: string;
  email: string;
  phone_number: string;
  shipping_status: string;
  tracking_number: string;
  label_url: string;
  estimated_delivery: string;
}